+++ projects/egeo/src/lib/st-switch/st-switch.model.ts
@@ -1,4 +1,4 @@
-/**
+/*
  * © 2017 Stratio Big Data Inc., Sucursal en España.
  *
  * This software is licensed under the Apache License, Version 2.0.
@@ -8,3 +8,8 @@
  *
  * SPDX-License-Identifier: Apache-2.0.
  */
+
+export enum StSwitchPosition {
+   LEFT = 'left',
+   RIGHT = 'right'
+}